pyproject.toml
[project]
name = "video-rag"
version = "0.1.0"
description = "Spatial event memory RAG system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.9",
    "fastapi>=0.135.1",
    "loguru>=0.7.3",
    "networkx>=3.6.1",
    "numpy>=2.4.3",
    "pydantic-settings>=2.13.1",
    "sentence-transformers>=5.3.0",
    "torch>=2.10.0",
    "uvicorn[standard]>=0.42.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
